# doc/swagger/apply.yaml
# Example YAML to get you started quickly.
# Be aware that YAML has indentation based scoping.
# Code completion support is available so start typing for available options.
swagger: '2.0'

# This is your document metadata
info:
  version: "1.0.0"
  title: 作者接口

#host: 127.0.0.1:7443
#basePath: /v1
#schemes:
#  - https
#produces:
#  - application/json
#consumes:
#  - application/json

# Describe your paths here
paths:
  /applys:
    get:
      summary: 查询全部报名信息
      description: |
        查询全部报名信息
      operationId: findAll
      responses:
        '200':
          description: Successful response
          schema:
            $ref: '#/definitions/responseModel'
        default:
          description: unexpected error
          schema:
            $ref: '#/definitions/errorModel'
      tags:
        - apply
    post:
      summary: 新增报名信息
      description: |
        新增报名信息
      operationId: saveApply
      produces:
        - application/json
      consumes:
        - application/json
      parameters:
        - in: body
          name: apply
          description: 报名信息
          schema:
            $ref: '#/definitions/apply'
      responses:
        '200':
          description: Successful response
          schema:
            $ref: '#/definitions/responseModel'
        default:
          description: unexpected error
          schema:
            $ref: '#/definitions/errorModel'
      tags:
        - apply
  /applys/{id}:
    get:
      summary: id查询
      description: |
        根据id查询报名信息
      operationId: getApply
      parameters:
        - name: id
          in: path
          description: id
          format: int64
          type: integer
          required: true
      responses:
        '200':
          description: Successful response
          schema:
            $ref: '#/definitions/responseModel'
        default:
          description: unexpected error
          schema:
            $ref: '#/definitions/errorModel'
      tags:
        - apply
    delete:
      tags:
        - apply
      summary: id删除
      description: |
        根据id删除报名信息
      operationId: deleteApply
      parameters:
        - name: id
          in: path
          description: id
          format: int64
          type: integer
          required: true
      responses:
        '200':
          description: Successful response
          schema:
            $ref: '#/definitions/responseModel'
        default:
          description: unexpected error
          schema:
            $ref: '#/definitions/errorModel'
    put:
      tags:
        - apply
      summary: 修改报名信息
      description: |
        根据id修改报名信息
      operationId: updateApply
      produces:
        - application/json
      consumes:
        - application/json
      parameters:
        - name: id
          in: path
          description: id
          format: int64
          type: integer
          required: true
        - in: body
          name: apply
          description: 报名信息
          schema:
            $ref: '#/definitions/apply'
      responses:
        '200':
          description: Successful response
          schema:
            $ref: '#/definitions/responseModel'
        default:
          description: unexpected error
          schema:
            $ref: '#/definitions/errorModel'
  /applys/page:
    get:
      tags:
        - apply
      summary: 项目名分页查询
      description: |
        根据项目名分页查询报名信息
      operationId: getByPage
      parameters:
        - name: itemName
          in: query
          description: 项目名
          type: string
        - name: start
          in: query
          description: 分页开始索引
          type: integer
          format: int32
        - name: length
          in: query
          description: 每页数量
          type: integer
          format: int32
      responses:
        '200':
          description: Successful response
          schema:
            $ref: '#/definitions/responseModel'
        default:
          description: unexpected error
          schema:
            $ref: '#/definitions/errorModel'
  /applys/phone/{phone}:
    get:
      tags:
        - apply
      summary: 手机号查询
      description: |
        根据手机号查询查询报名信息
      operationId: getByPhone
      parameters:
        - name: phone
          in: path
          description: 手机号
          type: string
          required: true
      responses:
        '200':
          description: Successful response
          schema:
            $ref: '#/definitions/responseModel'
        default:
          description: unexpected error
          schema:
            $ref: '#/definitions/errorModel'
definitions:
  apply:
    type: object
    properties:
      name:
        type: string
        maxLength: 20
        description: 姓名
      sex:
        type: string
        maxLength: 1
        description: 性别
      phone:
        type: string
        maxLength: 11
        description: 姓名
      itemName:
        type: string
        maxLength: 20
        description: 项目名
  responseModel:
    type: object
    required:
      - code
      - message
    properties:
      code:
        type: integer
        format: int32
      message:
        type: string
      data:
        type: string
  errorModel:
    type: object
    required:
      - code
      - message
    properties:
      code:
        type: integer
        format: int32
      message:
        type: string
# Added by API Auto Mocking Plugin
host: virtserver.swaggerhub.com
basePath: /garen.org/oss/1.0.0
schemes:
 - https
 - http